refactor[api-server]: 파츠 패키지 lottery 패키지에서 분리
