+++ src/main/java/com/watermelon/server/event/lottery/domain/LotteryApplier.java
@@ -1,8 +1,8 @@
 package com.watermelon.server.event.lottery.domain;
 
 import com.watermelon.server.BaseEntity;
-import com.watermelon.server.event.lottery.parts.domain.LotteryApplierParts;
-import com.watermelon.server.event.lottery.parts.exception.PartsDrawLimitExceededException;
+import com.watermelon.server.event.parts.domain.LotteryApplierParts;
+import com.watermelon.server.event.parts.exception.PartsDrawLimitExceededException;
 import jakarta.persistence.*;
 import lombok.*;
