+++ src/main/java/com/watermelon/server/event/lottery/service/LotteryServiceImpl.java
@@ -1,17 +1,9 @@
 package com.watermelon.server.event.lottery.service;
 
-import com.watermelon.server.admin.dto.response.ResponseAdminLotteryWinnerDto;
-import com.watermelon.server.admin.dto.response.ResponseAdminPartsWinnerDto;
 import com.watermelon.server.admin.dto.response.ResponseLotteryApplierDto;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.event.lottery.domain.LotteryReward;
-import com.watermelon.server.event.lottery.dto.request.RequestLotteryWinnerInfoDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryRankDto;
-import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerDto;
-import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerInfoDto;
-import com.watermelon.server.event.lottery.dto.response.ResponseRewardInfoDto;
-import com.watermelon.server.event.lottery.parts.domain.PartsReward;
-import com.watermelon.server.event.lottery.parts.repository.PartsRewardRepository;
 import com.watermelon.server.event.lottery.repository.LotteryApplierRepository;
 import com.watermelon.server.event.lottery.repository.LotteryRewardRepository;
 import jakarta.transaction.Transactional;
@@ -23,7 +15,6 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
