+++ src/main/java/com/watermelon/server/event/parts/domain/Parts.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.parts.domain;
+package com.watermelon.server.event.parts.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
