+++ src/main/java/com/watermelon/server/event/parts/exception/PartsNotExistException.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.parts.exception;
+package com.watermelon.server.event.parts.exception;
 
 public class PartsNotExistException extends RuntimeException {
 }
