+++ src/test/java/com/watermelon/server/ControllerTest.java
@@ -2,29 +2,25 @@ package com.watermelon.server;
 
 import com.epages.restdocs.apispec.ResourceSnippet;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
-import com.epages.restdocs.apispec.ResourceSnippetParametersBuilder;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.watermelon.server.admin.controller.LotteryEventService;
 import com.watermelon.server.admin.dto.response.ResponseAdminLotteryWinnerDto;
 import com.watermelon.server.admin.dto.response.ResponseAdminPartsWinnerDto;
 import com.watermelon.server.admin.dto.response.ResponseLotteryApplierDto;
-import com.watermelon.server.event.lottery.dto.request.RequestLotteryWinnerInfoDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryRankDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerInfoDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseRewardInfoDto;
 import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
 import com.watermelon.server.event.lottery.link.service.LinkService;
-import com.watermelon.server.event.lottery.parts.dto.response.ResponseMyPartsListDto;
-import com.watermelon.server.event.lottery.parts.dto.response.ResponsePartsDrawDto;
-import com.watermelon.server.event.lottery.parts.dto.response.ResponseRemainChanceDto;
-import com.watermelon.server.event.lottery.parts.exception.PartsDrawLimitExceededException;
-import com.watermelon.server.event.lottery.parts.service.PartsService;
+import com.watermelon.server.event.parts.dto.response.ResponseMyPartsListDto;
+import com.watermelon.server.event.parts.dto.response.ResponsePartsDrawDto;
+import com.watermelon.server.event.parts.dto.response.ResponseRemainChanceDto;
+import com.watermelon.server.event.parts.exception.PartsDrawLimitExceededException;
+import com.watermelon.server.event.parts.service.PartsService;
 import com.watermelon.server.event.lottery.service.LotteryRewardService;
 import com.watermelon.server.event.lottery.service.LotteryService;
 import com.watermelon.server.event.lottery.service.LotteryWinnerService;
 import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
@@ -32,19 +28,12 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
 
-import java.beans.ParameterDescriptor;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import static com.epages.restdocs.apispec.ResourceDocumentation.*;
 import static com.watermelon.server.Constants.*;
-import static com.watermelon.server.Constants.TEST_TOKEN;
 import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
