+++ src/test/java/com/watermelon/server/PartsRegistrationConfig.java
@@ -1,7 +1,7 @@
 package com.watermelon.server;
 
-import com.watermelon.server.event.lottery.parts.domain.Parts;
-import com.watermelon.server.event.lottery.parts.repository.PartsRepository;
+import com.watermelon.server.event.parts.domain.Parts;
+import com.watermelon.server.event.parts.repository.PartsRepository;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
