+++ src/test/java/com/watermelon/server/lottery/parts/service/LotteryApplierPartsServiceImplTest.java
@@ -1,12 +1,12 @@
 package com.watermelon.server.lottery.parts.service;
 
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
-import com.watermelon.server.event.lottery.parts.domain.LotteryApplierParts;
-import com.watermelon.server.event.lottery.parts.domain.Parts;
-import com.watermelon.server.event.lottery.parts.domain.PartsCategory;
-import com.watermelon.server.event.lottery.parts.repository.LotteryApplierPartsRepository;
-import com.watermelon.server.event.lottery.parts.repository.PartsRepository;
-import com.watermelon.server.event.lottery.parts.service.LotteryApplierPartsServiceImpl;
+import com.watermelon.server.event.parts.domain.LotteryApplierParts;
+import com.watermelon.server.event.parts.domain.Parts;
+import com.watermelon.server.event.parts.domain.PartsCategory;
+import com.watermelon.server.event.parts.repository.LotteryApplierPartsRepository;
+import com.watermelon.server.event.parts.repository.PartsRepository;
+import com.watermelon.server.event.parts.service.LotteryApplierPartsServiceImpl;
 import com.watermelon.server.event.lottery.service.LotteryApplierService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
