+++ src/test/java/com/watermelon/server/lottery/parts/service/PartsServiceImplTest.java
@@ -1,14 +1,14 @@
 package com.watermelon.server.lottery.parts.service;
 
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
-import com.watermelon.server.event.lottery.parts.domain.LotteryApplierParts;
-import com.watermelon.server.event.lottery.parts.domain.Parts;
-import com.watermelon.server.event.lottery.parts.domain.PartsCategory;
-import com.watermelon.server.event.lottery.parts.dto.response.ResponsePartsDrawDto;
-import com.watermelon.server.event.lottery.parts.exception.PartsDrawLimitExceededException;
-import com.watermelon.server.event.lottery.parts.repository.PartsRepository;
-import com.watermelon.server.event.lottery.parts.service.LotteryApplierPartsService;
-import com.watermelon.server.event.lottery.parts.service.PartsServiceImpl;
+import com.watermelon.server.event.parts.domain.LotteryApplierParts;
+import com.watermelon.server.event.parts.domain.Parts;
+import com.watermelon.server.event.parts.domain.PartsCategory;
+import com.watermelon.server.event.parts.dto.response.ResponsePartsDrawDto;
+import com.watermelon.server.event.parts.exception.PartsDrawLimitExceededException;
+import com.watermelon.server.event.parts.repository.PartsRepository;
+import com.watermelon.server.event.parts.service.LotteryApplierPartsService;
+import com.watermelon.server.event.parts.service.PartsServiceImpl;
 import com.watermelon.server.event.lottery.service.LotteryService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
